requirements.txt
gTTS==2.3.1
openai==0.26.4
playsound==1.2.2
python-dotenv==0.21.1
SpeechRecognition==3.9.0
